==> 1_Reddit_scraper.ipynb <==
import praw
import json
import datetime


reddit = praw.Reddit(client_id='IcgDFvvG5e6-D3wUvL6aow', client_secret='Nqzw-3Ab8jPEe9l3WNkZ_YhgJbllkw', user_agent='Data Scraping')

import pandas as pd
from pmaw import PushshiftAPI
api = PushshiftAPI()

import datetime as dt
before = int(dt.datetime(2021,6,30,0,0).timestamp())
after = int(dt.datetime(2021,1,1,0,0).timestamp())

subreddit="ethereum"
limit=1000000
comments = api.search_comments(subreddit=subreddit, limit=limit, before=before, after=after)
print(f'Retrieved {len(comments)} comments from Pushshift')

api.metadata_.get('shards')


comments_df = pd.DataFrame(comments)
# preview the comments data
comments_df.head(5)

comments_df.columns

comments_df['collapsed'].value_counts()

len(comments_df)

comments_df['controversiality'].head()

ethereum_df = comments_df[['created_utc', 'author', 'author_fullname','body', 'subreddit','subreddit_id', 'send_replies', 'no_follow']]

len(ethereum_df)

ethereum_df['DateTime'] = pd.to_datetime(ethereum_df['created_utc'], unit = 's')
ethereum_df['Dates'] = pd.to_datetime(ethereum_df['DateTime']).dt.date
ethereum_df['Time'] = pd.to_datetime(ethereum_df['DateTime']).dt.time

ethereum_df.head()

ethereum_df.to_csv("ethereum_subreddits.csv")

subreddit="CryptoCurrency"
limit=1000000
comments_c = api.search_comments(subreddit=subreddit, limit=limit, before=before, after=after)
print(f'Retrieved {len(comments)} comments from Pushshift')

api.metadata_.get('shards')

crypto = pd.DataFrame(comments_c)
# preview the comments data
crypto.head(5)

crypto_df = crypto[['created_utc', 'author', 'author_fullname','body', 'subreddit','subreddit_id', 'send_replies', 'no_follow']]

crypto_df.head()

crypto_df['DateTime'] = pd.to_datetime(crypto_df['created_utc'], unit = 's')

crypto_df.head()

crypto_df.to_csv("crypto.csv")

crypto_df['Dates'] = pd.to_datetime(crypto_df['DateTime']).dt.date
crypto_df['Time'] = pd.to_datetime(crypto_df['DateTime']).dt.time

crypto_df.head()
