==> 4_format_for_analysis.ipynb <==
# Data Preparation for Analysis

# import packages
import pandas as pd
import numpy as np
import datetime as dt
import matplotlib.pyplot as plt

### sentiment data

# import data
data = pd.read_csv('sentiment_data_odd.csv', index_col = 0)

data.head()

# concatencate even and odd sentiment_data into one df


# ['created_utc'] of head should be even, tail should be odd
# even data is on top, append means you stick data to the bottom so odd is on the bottom
data.head(), data.tail()

# delete redundant data
del even
del odd

# convert ['created_utc'] to datetime date
data['datetime'] = pd.to_datetime(data['created_utc'], unit = 's')
data.head()

# convert ['datetime'] to date
data['date'] = pd.DatetimeIndex(data['datetime']).date
data.head()

data['date']=pd.to_datetime(data['date'])

# delete redundant data
del data['created_utc']
del data['datetime']

# Here you can create new variables/ group_by different variables. 
# I don't use ['score'] and ['subreddit'] in this analysis but there are different ways that you could possibly incorporate them into your analysis
# for example: using ['score'] to assign weights to comments, cryptocurrency vs non-cryptocurrency ['subreddit'] sentiment, etc....

# for now i will delete them

data.head()


# groupby date
# returns count of positive/neutral/negative comments for each ['date']
# ['date'] is now the index
data = data.groupby(by = ['date']).sum()
data.head()
del data['send_replies']
del data['no_follow']
del data['neg']
del data['neu']
del data['pos']
del data['compound']

data.head()

## price data

price = pd.read_csv('ETH-USD.csv')
price.head()
type(price['Date'])
price['Date']=pd.to_datetime(price['Date'])


# create new date variable that is the same name as the date variable in `data
price['date'] = price['Date'].copy()

# check to see if `price` and `data` have the same `date` format

print('-- string --')
print('data : ' + str(data.index[0]))
print('price : ' + str(price['date'][0]))
print('------------------------------')
print('-- raw string --')
print('data : ' + repr(data.index[0]))
print('price : ' + repr(price['date'][0]))

While the strings may look the same, they are actually different. `data['date']` is actually<br>
a `datetime.date` object type while `price['date']` is just a string. We see this when we <br>
print the raw strings, which is what the data actually is and the data that python interprets.<br>
In this state, they can't merge. We have to convert `price['date']` to a `datetime.date` object. 

# convert to `datetime` object
price['date'] = pd.to_datetime(price['date'])
str(price['date'][0]), price['date'][0]

# convert datetime into `datetime.date` object
price['date'] = pd.DatetimeIndex(price['date']).date
str(price['date'][0]), price['date'][0]

# check to see if `price` and `data` have the same `date` format

print('-- string --')
print('data : ' + str(data.index[0]))
print('price : ' + str(price['Date'][0]))
print('------------------------------')
print('-- raw string --')
print('data : ' + repr(data.index[0]))
print('price : ' + repr(price['Date'][0]))

# define `price` index as `price['date']`
price.index = price['Date']
del price['Date']
price.head()

## merging data and price

# combine sentiment and price data
data = price.merge(data, how = 'left', left_index = True, right_index = True)
data.head()

# don't worry about the NaN values. That just means that on that day, there weren't any comments as the bitcoin community was still small

# last date that we have sentiment data for
end_date = dt.date(2021, 6, 30)

# redefine data with `end_date` as the last point
data = data[ : end_date]
data.tail()


# new df with formatted variables
df = pd.DataFrame()
df['price'] = data['Close']
df['volatility'] = (data['High'] - data['Low']) / data['Close']
df['volume_price'] = data['Volume']
df['volume_number'] = data['Volume'] / data['Close']
df['positive_comment'] = data['positive_comment']
df['neutral_comment'] = data['neutral_comment']
df['negative_comment'] = data['negative_comment']

df.tail()

df.to_csv('formatted_data_all.csv')
