==> 5_sentiment_analysis.ipynb <==
# Sentiment Analysis

# import packages
import pandas as pd
import multiprocessing as mp
import tqdm
from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer

# import data
data = pd.read_csv('clean_data.csv', index_col = 0)
data.head()
type(data['body'])
data["body"] = data["body"].astype(str)
print(type(data['body']))
data['body'].to_string()
print(type(data['body']))

```Python
# initialize sentiment classifier
sid = sid()

# get sentiment scores from data
sentiment = data['body'].apply(SID.polarity_scores)

# convert sentiment series into dataframe (each sentiment value gets its own column)
sentiment = pd.DataFrame(sentiment.tolist())

# merge data and sentiment df into one
data = data.merge(sentiment, how = 'left', left_index = True, right_index = True)

# delete sentiment as its info is in data
del sentiment
```

# initialize sentiment classifier
sid = SentimentIntensityAnalyzer()

# get sentiment scores from data
sentiment = data['body'].apply(sid.polarity_scores)

# convert sentiment series into dataframe (each sentiment value gets its own column)
sentiment = pd.DataFrame(sentiment.tolist())

# merge data and sentiment df into one
data = data.merge(sentiment, how = 'left', left_index = True, right_index = True)

# delete sentiment as its info is in data
del sentiment

data.head()

data.head()

## categorize sentiment data into sentiment categories

# create function to categorize compound sentiment score
# 0.05 threshold recommended on VADER documentation
# you should read the comments, read its compound score, and determine your own cuttoffs
def categorize_sentiment(x):
    if x >= 0.05:
        return 'positive_comment'
    elif 0.05 > x > -0.05:
        return 'neutral_comment'
    elif -0.05 >= x:
        return 'negative_comment'

# apply function categorize_sentiment to ['compound']
data['sentiment'] = data['compound'].apply(categorize_sentiment)

data.head()

# convert ['sentiment'] to categorical data type
data['sentiment'] = pd.Categorical(data['sentiment'])

#sentiment should be category
data.dtypes

# convert ['sentiment'] categories to binary variables in new df
binary_sentiment = data['sentiment'].str.get_dummies()
binary_sentiment.head()

# count of how many of each category were classified
binary_sentiment.sum()

# merge binary_sentiment with data 
data = data.merge(binary_sentiment, how = 'left', left_index = True, right_index = True)
data.head()

# delete redundant variables
del data['pos']
del data['neg']
del data['neu']
del data['compound']
del data['body']
del data['sentiment']
del binary_sentiment

# export data
data.to_csv('sentiment_data.csv')
